fix(benchmarks): stabilize read sampling evidence
