+++ benchmarks/README.md
@@ -14,9 +14,30 @@ Refresh the smoke baseline only when a performance change is intentionally accep
 vp run -w bench:baseline:smoke:update
 ```
 
+That updater replaces every task in the profile. For an accepted measurement-protocol migration,
+run and refresh only the affected tasks so unrelated reference task values remain unchanged:
+
+```bash
+node scripts/run-benchmark-baseline.mjs --profile=smoke --update-baseline \
+  '--update-baseline-task=raw snapshot 1000 rows' \
+  '--update-baseline-task=raw predicate index 1000 rows' \
+  '--update-baseline-task=grouped aggregate 1000 rows' \
+  '--update-baseline-task=grouped key width 1000 rows'
+```
+
+Scoped updates require the committed and current profile task catalogs to match. They may refresh
+only timing samples, RSS, minimum sample counts, and sampling-policy metadata for selected tasks;
+workload identity or structural metadata drift aborts without rewriting the baseline.
+
 The smoke gate runs the existing Vitest benchmark files serially and compares fresh `.artifacts`
 summary/output JSON against the committed baseline. Engine smoke cases use small row counts with
-multiple samples; browser smoke stays deliberately tiny to keep CI practical. The gate fails on
+multiple samples; browser smoke stays deliberately tiny to keep CI practical. Read-focused raw
+snapshot, raw predicate, grouped aggregate, and grouped key-width cases use 1,000 minimum measured
+iterations, a 250 ms measurement floor, five warmup iterations, and a 100 ms warmup floor. The
+benchmark summary and committed baseline retain this policy, and the runner rejects fewer samples
+or policy drift. The live-delta cases in the affected mixed raw-snapshot and grouped-aggregate tasks
+retain exactly five iteration-bound samples with time and warmup disabled; fanout and other mutation
+tasks keep their existing sampling policy. The gate fails on
 cleanup leaks, backpressure, queued-event growth, RSS growth, mean latency regressions, or p99
 latency regressions beyond the code-owned thresholds mirrored in the baseline manifest. Latency
 thresholds intentionally use the wider of the ratio and absolute windows because CI smoke runs are
@@ -63,15 +84,32 @@ Refresh it only when a raw read/write performance change is intentionally accept
 vp run -w bench:baseline:raw-read-write:update
 ```
 
+For a read-path measurement-protocol migration, preserve both write tasks and update only the
+affected read observations:
+
+```bash
+node scripts/run-benchmark-baseline.mjs --profile=raw-read-write --update-baseline \
+  '--update-baseline-task=raw snapshot 100000 rows' \
+  '--update-baseline-task=raw predicate index 100000 rows'
+```
+
 This profile is localhost CPU/GC engine stress. It runs 100k-row raw snapshots, predicate-index
 reads, base writes, and indexed writes. It exists to catch the common failure mode where a read-path
 optimization wins filtered/sorted queries but silently taxes ingestion. Keep this gate serial and do
 not compare it against runs collected while another benchmark is active.
-Raw write cases run with exact sample and mutation counts. The profile keeps mean latency tighter
+Read cases use 1,000 minimum measured iterations, a 250 ms measurement floor, five warmup iterations,
+and a 100 ms warmup floor. The raw snapshot live-delta case and raw writes keep time and warmup
+disabled and run with exactly 20 samples and matching mutation counts. The profile keeps mean latency tighter
 than smoke for millisecond-scale work, but keeps a small absolute window for sub-millisecond cases so
-GitHub runner jitter does not fail healthy 0.xms operations. p99 remains a wide tail-noise guard
-because with exact local samples it behaves like max latency and is sensitive to scheduler/GC spikes;
-use repeated local runs before treating p99-only movement in this profile as a product regression.
+GitHub runner jitter does not fail healthy 0.xms operations. p99 remains a wide tail-noise guard.
+The 1,000-sample read population leaves roughly ten observations above the p99 rank, so two or three
+scheduler/GC pauses cannot both stop sampling at its minimum and define p99. The runner retains every
+sample and fails the first comparison; it does not trim outliers, retry, or select a best run.
+Affected read tasks measure RSS as the process-lifetime peak reached through setup, warmup, JIT, GC,
+and measured work, anchored against the initial current RSS in the fresh benchmark worker. This
+conservative metric cannot be lowered by a lucky final garbage collection. The endpoint memory
+snapshots remain in the summary for diagnosis, while the runner compares only the policy-owned peak
+delta with the existing RSS thresholds.
 
 WebSocket firehose smoke has a focused runtime transport gate:
