+++ package.json
@@ -44,7 +44,7 @@
     "release:version": "vp exec changeset version",
     "release-candidate:capacity": "vp run -w examples:test && vp run -w examples:build && vp run -w pre-grpc:gate && vp run -w grpc:gate && vp run -w bench:baseline:release",
     "test:benchmark-baseline": "vp run -w test:repository-scripts",
-    "test:repository-scripts": "vp test run scripts/benchmark-artifact-mechanics.test.ts scripts/benchmark-baseline.test.ts scripts/grpc-materialized-benchmark-policy.test.ts scripts/grpc-leased-benchmark-policy.test.ts scripts/benchmark-baseline-runner.test.ts scripts/bench-runtime-kafka-ingest.test.ts scripts/package-surface-policy.test.ts scripts/typescript-module-inspection.test.ts scripts/check-internal-seams.test.ts scripts/package-export-script.test.ts scripts/release-publish-policy.test.ts scripts/release-publish-orchestration.test.ts --coverage",
+    "test:repository-scripts": "vp test run scripts/benchmark-artifact-mechanics.test.ts scripts/benchmark-baseline.test.ts scripts/benchmark-baseline-sampling-policy.test.ts scripts/benchmark-sampling-policy.test.ts scripts/grpc-materialized-benchmark-policy.test.ts scripts/grpc-leased-benchmark-policy.test.ts scripts/benchmark-baseline-runner-profile-policy.test.ts scripts/benchmark-baseline-runner-profile-catalog.test.ts scripts/benchmark-baseline-runner-baseline-update.test.ts scripts/benchmark-baseline-runner-execution.test.ts scripts/benchmark-baseline-cli.test.ts scripts/bench-runtime-kafka-ingest.test.ts scripts/package-surface-policy.test.ts scripts/typescript-module-inspection.test.ts scripts/check-internal-seams.test.ts scripts/package-export-script.test.ts scripts/release-publish-policy.test.ts scripts/release-publish-orchestration.test.ts --coverage",
     "prepare": "vp config && effect-language-service patch"
   },
   "devDependencies": {
