+++ packages/column-live-view-engine/src/raw-active-retained-delta.bench.ts
@@ -16,11 +16,10 @@ import {
   cleanupLeakCountFromEngineHealth,
   failOnBenchmarkCleanupLeaks,
   isBenchmarkEngineHealth,
-  memorySnapshot,
   queuedEventCountFromEngineHealth,
   writeBenchmarkArtifact,
-  type BenchmarkMemorySnapshot,
 } from "./benchmark-artifact";
+import { memorySnapshot, type BenchmarkMemorySnapshot } from "./benchmark-memory-recorder";
 
 declare const process: {
   readonly env: Record<string, string | undefined>;
