+++ vite.config.ts
@@ -9,7 +9,10 @@ export default defineConfig({
         "scripts/benchmark-artifact-mechanics.mjs",
         "scripts/benchmark-baseline.mjs",
         "scripts/benchmark-baseline-cli.mjs",
+        "scripts/benchmark-baseline-profiles.mjs",
         "scripts/benchmark-baseline-runner.mjs",
+        "scripts/benchmark-baseline-task-catalog.mjs",
+        "scripts/benchmark-sampling-policy.mjs",
         "scripts/bench-runtime-kafka-ingest.mjs",
         "scripts/check-internal-seams.ts",
         "scripts/grpc-leased-benchmark-policy.mjs",
